Fix 6 bugs found during live testing
- chunker: stat-poll init.mp4 after inotify fires so initSegmentPath is only
  set once the file has content; previously the creation event fired before
  ffmpeg finished writing, sending a 0-byte init segment that corrupted the
  MSE SourceBuffer and broke all streaming
- schema: change fileSizeBytes Int! → Float! so 4K movie files (>2.1 GB) no
  longer overflow GraphQL's 32-bit int and crash the Profiles page
- formatters: resolutionLabel/maxResolutionForHeight now accept an optional
  width param; width >= 3840 maps to "4k" so Cinemascope 4K sources
  (3840×1600, 3840×1608) are no longer misclassified as 1080p
- Add width to videoStream fragment in all 7 components (VideoCard,
  MediaListItem, MediaGridItem, VideoDetailsPanel, ControlBar, VideoPlayer,
  PlayerSidebar); update call sites to pass width; regenerate Relay artifacts
- PlayerSidebar: remove .toUpperCase() on resolution badge ("1080P" → "1080p")
- stream.ts: replace hardcoded Access-Control-Allow-Origin: localhost:5173
  with wildcard so dev CORS works regardless of which port Vite picks
- VideoPlayer: make big play overlay clickable (remove pointerEvents:none,
  wire onClick to handlePlay) so users can start playback without waiting
  for the auto-hide ControlBar to become visible

+++ server/src/graphql/schema.ts
@@ -33,7 +33,7 @@ export const typeDefs = /* GraphQL */ `
     title: String!
     filename: String!
     durationSeconds: Float!
-    fileSizeBytes: Int!
+    fileSizeBytes: Float!
     bitrate: Int!
     library: Library!
     videoStream: VideoStreamInfo
